# Add multithreading to ordered dithering