@adrhill 2022-09-03
Looks like benchmarks are broken on master, so here are some local benchmarks:
### Master `v3.0.3`
```Julia
julia> using DitherPunk, TestImages, ColorSchemes

julia> img = rand(Gray, 1000, 1000);

julia> imc = testimage("fabio_color_256");

julia> alg = Bayer(3);

julia> cs = ColorSchemes.jet;

julia> @benchmark dither($img, $alg)

BenchmarkTools.Trial: 2574 samples with 1 evaluation.
 Range (min … max):  745.018 μs … 17.749 ms  ┊ GC (min … max):  0.00% … 91.93%
 Time  (median):       1.203 ms              ┊ GC (median):     0.00%
 Time  (mean ± σ):     1.942 ms ±  2.794 ms  ┊ GC (mean ± σ):  38.72% ± 22.87%

  ▂▇█                                                           
  ████▄▄▁▃▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃▃▅▅▅▅▅▄▆▆▇▆▇▅▆▇▆▇▆▆▇▅▅▆▄▄▅ █
  745 μs        Histogram: log(frequency) by time        14 ms <

 Memory estimate: 7.65 MiB, allocs estimate: 5.

julia> @benchmark dither($imc, $alg, $cs)
BenchmarkTools.Trial: 8 samples with 1 evaluation.
 Range (min … max):  622.878 ms … 683.338 ms  ┊ GC (min … max): 2.68% … 3.48%
 Time  (median):     639.255 ms               ┊ GC (median):    4.63%
 Time  (mean ± σ):   642.701 ms ±  20.433 ms  ┊ GC (mean ± σ):  4.45% ± 1.52%

  ███            █ █            ██                            █  
  ███▁▁▁▁▁▁▁▁▁▁▁▁█▁█▁▁▁▁▁▁▁▁▁▁▁▁██▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█ ▁
  623 ms           Histogram: frequency by time          683 ms <

 Memory estimate: 155.58 MiB, allocs estimate: 331455.

```
### This PR
```Julia
ulia> @benchmark dither($img, $alg)
BenchmarkTools.Trial: 2403 samples with 1 evaluation.
 Range (min … max):  666.558 μs … 29.729 ms  ┊ GC (min … max):  0.00% … 97.15%
 Time  (median):     756.130 μs              ┊ GC (median):     0.00%
 Time  (mean ± σ):     2.071 ms ±  4.782 ms  ┊ GC (mean ± σ):  59.81% ± 23.67%

  █▄                                                            
  ███▅▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃▁▁▁▁▁▃▁▃▃▄▄▅▃▅▄▆▆▁▆▅▅▆▆▆▅▆▆▆▆▅▅▅▆▅▅▅▁▆ █
  667 μs        Histogram: log(frequency) by time        24 ms <

 Memory estimate: 7.65 MiB, allocs estimate: 54.

julia> @benchmark dither($imc, $alg, $cs)
BenchmarkTools.Trial: 26 samples with 1 evaluation.
 Range (min … max):  103.469 ms … 423.564 ms  ┊ GC (min … max):  0.00% … 69.35%
 Time  (median):     118.208 ms               ┊ GC (median):     0.00%
 Time  (mean ± σ):   195.786 ms ± 115.151 ms  ┊ GC (mean ± σ):  41.62% ± 32.75%

  ▃▁█▁                                                           
  ████▁▁▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄▄▄▄▁▁▄▁▁▄▁▇▁▁▁▁▁▁▁▁▁▄ ▁
  103 ms           Histogram: frequency by time          424 ms <

```