@adrhill 2022-09-03
Looks like the performance increase on grayscale images isn't as clear-cut as I thought when compared to `@simd`.