@adrhill 2023-01-15
Closing this since performance increases weren't clear-cut in the first place and `@simd` seems to be the better option.